Optimized Solution Added

+++ UserInput.cs
@@ -4,15 +4,15 @@ namespace NP_CompleteOrder
 {
     public class UserInput
     {
-        public static double GetSpendTotal(double defaultSpendTotal)
+        public static decimal GetSpendTotal(decimal defaultSpendTotal)
         {
-            double spendTotal = 0;
+            decimal spendTotal = 0;
             string userInput = null;
             do
             {
                 Console.Write($"How much would you like to spend? [${string.Format("{0:#.00}", defaultSpendTotal)}]:");
                 userInput = Console.ReadLine();
-            } while (userInput != "" && !double.TryParse(userInput, out spendTotal));
+            } while (userInput != "" && !decimal.TryParse(userInput, out spendTotal));
 
             if (spendTotal == 0)
             {
